fixed epollerr loop non post cgi, crash errorcodeClient resolved, python edit for get request, remove post handle when error occurs within

+++ include/ErrorCodeClientException.hpp
@@ -10,7 +10,7 @@ class ErrorCodeClientException
 {
     public:
         // Initialization
-        explicit ErrorCodeClientException(Client &client, int errorCode, const string &message);
+        explicit ErrorCodeClientException(Client &client, uint16_t errorCode, const string &message);
 
         //handle error page
         void handleErrorClient() const;
